requirements.txt
numpy<2.0.0
pandas==2.2.3
pika==1.3.1
prophet==1.1.5
psycopg2-binary==2.9.9
python-dotenv==1.1.0
Requests==2.32.3
scikit_learn==1.6.1
torch==2.5.1
